Removing some miner files and some refactoring
