+++ txs/src/lib.rs
@@ -16,16 +16,10 @@
 )]
 
 pub mod application;
-pub mod block;
 pub mod commands;
 pub mod config;
-pub mod delay;
 pub mod error;
 pub mod prelude;
 pub mod submit_tx;
-pub mod test_tx_swarm;
-pub mod backlog;
 // pub mod node_keys;
 pub mod keygen;
-pub mod account;
-
