+++ txs/src/submit_tx.rs
@@ -15,8 +15,8 @@ use std::{io::{stdout, Write}, thread, time};
 
 use libra_types::transaction::{Script, TransactionArgument, TransactionPayload};
 use libra_types::{transaction::helpers::*};
-use crate::config::MinerConfig;
-use compiled_stdlib::transaction_scripts;
+// use crate::config::MinerConfig;
+// use compiled_stdlib::transaction_scripts;
 use libra_json_rpc_types::views::{TransactionView, VMStatusView};
 use libra_types::chain_id::ChainId;
 use libra_genesis_tool::keyscheme::KeyScheme;
@@ -96,173 +96,44 @@ pub fn submit_tx_(
 
 }
 
+// use crate::submit_tx::{TxParams};
+use crate::keygen;
+// use anyhow::Error;
+use libra_config::config::NodeConfig;
+// use libra_crypto::test_utils::KeyPair;
+// use reqwest::Url;
+use std::{path::PathBuf};
+// use libra_genesis_tool::keyscheme::KeyScheme;
 
-/// Submit a miner transaction to the network.
-pub fn submit_tx(
-    tx_params: &TxParams,
-    preimage: Vec<u8>,
-    proof: Vec<u8>,
-    is_onboading: bool,
-) -> Result<TransactionView, Error> {
-
-    // Create a client object
-    let mut client = LibraClient::new(tx_params.url.clone(), tx_params.waypoint).unwrap();
-
-    let chain_id = ChainId::new(client.get_metadata().unwrap().chain_id);
-
-    let (account_state,_) = client.get_account(tx_params.address.clone(), true).unwrap();
-    let sequence_number = match account_state {
-        Some(av) => av.sequence_number,
-        None => 0,
-    };
-
-    let script: Script;
-    // Create the unsigned MinerState transaction script
-    if !is_onboading {
-        script = Script::new(
-            transaction_scripts::StdlibScript::MinerStateCommit.compiled_bytes().into_vec(),
-            vec![],
-            vec![
-                TransactionArgument::U8Vector(preimage),
-                TransactionArgument::U8Vector(proof),
-            ],
-        );
-    } else {
-
-        let consensus_pubkey = hex::decode("8108aedfacf5cf1d73c67b6936397ba5fa72817f1b5aab94658238ddcdc08010").unwrap();
-        let validator_network_address = "test".as_bytes().to_vec();
-        let full_node_network_address = "test".as_bytes().to_vec();
-        let human_name = "test".as_bytes().to_vec();
-        script = Script::new(
-            transaction_scripts::StdlibScript::MinerStateOnboarding.compiled_bytes().into_vec(),
-            vec![],
-            vec![
-                TransactionArgument::U8Vector(preimage), // challenge: vector<u8>,
-                TransactionArgument::U8Vector(proof), // solution: vector<u8>,
-                TransactionArgument::U8Vector(consensus_pubkey), // consensus_pubkey: vector<u8>,
-                TransactionArgument::U8Vector(validator_network_address),// validator_network_address: vector<u8>,
-                TransactionArgument::U8Vector(full_node_network_address),// full_node_network_address: vector<u8>,
-                TransactionArgument::U8Vector(human_name),// human_name: vector<u8>,
-            ],
-        );
-    }
-
-    // sign the transaction script
-    let txn = create_user_txn(
-        &tx_params.keypair,
-        TransactionPayload::Script(script),
-        tx_params.address,
-        sequence_number,
-        tx_params.max_gas_unit_for_tx,
-        tx_params.coin_price_per_unit,
-        "GAS".parse()?,
-        tx_params.user_tx_timeout as i64, // for compatibility with UTC's timestamp.
-        chain_id,
-    )?;
-
-    // get account_data struct
-    let mut sender_account_data = AccountData {
-        address: tx_params.address,
-        authentication_key: Some(tx_params.auth_key.to_vec()),
-        key_pair: Some(tx_params.keypair.clone()),
-        sequence_number,
-        status: AccountStatus::Persisted,
-    };
-    
-    // Submit the transaction with libra_client
-    match client.submit_transaction(
-        Some(&mut sender_account_data),
-        txn
-    ){
-        Ok(_) => {
-            match wait_for_tx(tx_params.address, sequence_number, &mut client) {
-                Some(res) => Ok(res),
-                None => Err(Error::msg("No Transaction View returned"))
-            }
-        }
-        Err(err) => Err(err)
-    }
-
-}
-
-
-/// Submit a miner transaction to the network.
-pub fn submit_onboard_tx(
-    tx_params: &TxParams,
-    preimage: Vec<u8>,
-    proof: Vec<u8>,
-    ow_human_name: Vec<u8>,
-    op_address: AccountAddress,
-    op_auth_key_prefix: Vec<u8>,
-    op_consensus_pubkey: Vec<u8>,
-    op_validator_network_addresses: Vec<u8>,
-    op_fullnode_network_addresses: Vec<u8>,
-    op_human_name: Vec<u8>,
-) -> Result<TransactionView, Error> {
-
-    // Create a client object
-    let mut client = LibraClient::new(tx_params.url.clone(), tx_params.waypoint).unwrap();
-
-    let chain_id = ChainId::new(client.get_metadata().unwrap().chain_id);
-
-    let (account_state,_) = client.get_account(tx_params.address.clone(), true).unwrap();
-    let sequence_number = match account_state {
-        Some(av) => av.sequence_number,
-        None => 0,
-    };
-
-    let script = transaction_builder::encode_minerstate_onboarding_script(
-        preimage,
-        proof,
-        ow_human_name,
-        op_address,
-        op_auth_key_prefix,
-        op_consensus_pubkey,
-        op_validator_network_addresses,
-        op_fullnode_network_addresses,
-        op_human_name,
-        // vec!(),
-        // vec!(),
+/// Helper to extract params from a local running swarm.
+pub fn get_params_from_swarm(mut swarm_path: PathBuf) -> Result<TxParams, Error> {
+    swarm_path.push("0/node.yaml");
+    let config = NodeConfig::load(&swarm_path).unwrap_or_else(
+        |_| panic!("Failed to load NodeConfig from file: {:?}", &swarm_path)
     );
 
-    // sign the transaction script
-    let txn = create_user_txn(
-        &tx_params.keypair,
-        TransactionPayload::Script(script),
-        tx_params.address,
-        sequence_number,
-        tx_params.max_gas_unit_for_tx,
-        tx_params.coin_price_per_unit,
-        "GAS".parse()?,
-        tx_params.user_tx_timeout as i64, // for compatibility with UTC's timestamp.
-        chain_id,
-    )?;
+    let url =  Url::parse(
+        format!("http://localhost:{}", config.json_rpc.address.port()).as_str()
+    ).unwrap();
+    let waypoint = config.base.waypoint.genesis_waypoint();
 
-    // get account_data struct
-    let mut sender_account_data = AccountData {
-        address: tx_params.address,
-        authentication_key: Some(tx_params.auth_key.to_vec()),
-        key_pair: Some(tx_params.keypair.clone()),
-        sequence_number,
-        status: AccountStatus::Persisted,
+    let (auth_key, address, wallet) = keygen::account_from_prompt();
+    let keys = KeyScheme::new_from_mnemonic(wallet.mnemonic());
+    let keypair = KeyPair::from(keys.child_0_owner.get_private_key());
+
+    let tx_params = TxParams {
+        auth_key,
+        address,
+        url,
+        waypoint,
+        keypair,
+        max_gas_unit_for_tx: 1_000_000,
+        coin_price_per_unit: 1, // in micro_gas
+        user_tx_timeout: 5_000,
     };
-    
-    // Submit the transaction with libra_client
-    match client.submit_transaction(
-        Some(&mut sender_account_data),
-        txn
-    ){
-        Ok(_) => {
-            match wait_for_tx(tx_params.address, sequence_number, &mut client) {
-                Some(res) => Ok(res),
-                None => Err(Error::msg("No Transaction View returned"))
-            }
-        }
-        Err(err) => Err(err)
-    }
 
+    Ok(tx_params)
 }
-
 
 /// Wait for the response from the libra RPC.
 pub fn wait_for_tx(
@@ -309,98 +180,6 @@ pub fn eval_tx_status(result: TransactionView) -> bool {
     }
 }
 
-/// Form tx parameters struct 
-pub fn get_params(
-    keys: KeyScheme, 
-    waypoint: Waypoint,
-    config: &MinerConfig
-) -> TxParams {
-    // let keys = KeyScheme::new_from_mnemonic(mnemonic.to_string());
-    let keypair = KeyPair::from(keys.child_0_owner.get_private_key());
-    let pubkey =  &keypair.public_key;// keys.child_0_owner.get_public();
-    let auth_key = AuthenticationKey::ed25519(pubkey);
-    let url_str = config.chain_info.node.as_ref().unwrap();
-
-    TxParams {
-        auth_key,
-        address: config.profile.account,
-        url: Url::parse(url_str).expect("No url provided in miner.toml"),
-        waypoint,
-        keypair,
-        max_gas_unit_for_tx: 5_000,
-        coin_price_per_unit: 1, // in micro_gas
-        user_tx_timeout: 5_000,
-    }
-}
-
-
-/// Submit a miner transaction to the network.
-pub fn util_save_tx(
-    tx_params: &TxParams,
-){
-
-    // Create a client object
-    // let mut client = LibraClient::new(tx_params.url.clone(), tx_params.waypoint).unwrap();
-
-    let chain_id = ChainId::new(1);
-
-    // let (account_state,_) = client.get_account(tx_params.address.clone(), true).unwrap();
-    // let sequence_number = match account_state {
-    //     Some(av) => av.sequence_number,
-    //     None => 0,
-    // };
-
-    let script = transaction_builder::encode_demo_e2e_script(42);
-
-    // TODO, how does Alice get Bob's tx sequence number?
-    // sign the transaction script
-    let txn = create_user_txn(
-        &tx_params.keypair,
-        TransactionPayload::Script(script),
-        tx_params.address,
-        1,
-        tx_params.max_gas_unit_for_tx,
-        tx_params.coin_price_per_unit,
-        "GAS".parse().unwrap(),
-        tx_params.user_tx_timeout as i64, // for compatibility with UTC's timestamp.
-        chain_id,
-    );
-
-    match txn {
-        Ok(signed_tx) => {
-            println!("Signed tx: {:?}", signed_tx);
-        }
-        Err(e) => {
-            println!("Could not write tx: {:?}", e);
-        }
-    }
-
-    // // get account_data struct
-    // let mut sender_account_data = AccountData {
-    //     address: tx_params.address,
-    //     authentication_key: Some(tx_params.auth_key.to_vec()),
-    //     key_pair: Some(tx_params.keypair.clone()),
-    //     sequence_number,
-    //     status: AccountStatus::Persisted,
-    // };
-    
-    // // Submit the transaction with libra_client
-    // match client.submit_transaction(
-    //     Some(&mut sender_account_data),
-    //     txn
-    // ){
-    //     Ok(_) => {
-    //         match wait_for_tx(tx_params.address, sequence_number, &mut client) {
-    //             Some(res) => Ok(res),
-    //             None => Err(Error::msg("No Transaction View returned"))
-    //         }
-    //     }
-    //     Err(err) => Err(err)
-    // }
-
-}
-
-
 #[test]
 fn test_make_params() {
     use libra_types::PeerId; 
@@ -438,40 +217,4 @@ fn test_make_params() {
     assert_eq!("http://localhost:8080/".to_string(), p.url.to_string());
     // debug!("{:?}", p.url);
     //make_params
-}
-
-#[test]
-fn test_save_tx() {
-    use libra_types::PeerId; 
-    use crate::config::{
-        Workspace,
-        Profile,
-        ChainInfo
-    };
-    use std::path::PathBuf;
-
-    let mnemonic = "talent sunset lizard pill fame nuclear spy noodle basket okay critic grow sleep legend hurry pitch blanket clerk impose rough degree sock insane purse";
-    let waypoint: Waypoint =  "0:3e4629ba1e63114b59a161e89ad4a083b3a31b5fd59e39757c493e96398e4df2".parse().unwrap();
-    let configs_fixture = MinerConfig {
-        workspace: Workspace{
-            node_home: PathBuf::from("."),
-        },
-        profile: Profile {
-            auth_key: "3e4629ba1e63114b59a161e89ad4a083b3a31b5fd59e39757c493e96398e4df2"
-                .to_owned(),
-            account: PeerId::from_hex_literal("0x000000000000000000000000deadbeef").unwrap(),
-            ip: "1.1.1.1".parse().unwrap(),
-            statement: "Protests rage across the nation".to_owned(),
-        },
-        chain_info: ChainInfo {
-            chain_id: "0L testnet".to_owned(),
-            block_dir: "test_blocks_temp_2".to_owned(),
-            base_waypoint: None,
-            node: Some("http://localhost:8080".to_string()),
-        },
-
-    };
-    let keys = KeyScheme::new_from_mnemonic(mnemonic.to_owned());
-    let p = get_params(keys, waypoint, &configs_fixture);
-    util_save_tx(&p);
 }
